+++ anyblok/__init__.py
@@ -13,7 +13,8 @@ PROMPT = "%(processName)s - %(version)s"
 
 
 def start(processName, version=release.version, prompt=PROMPT,
-          argsparse_groups=None, parts_to_load=None, logger=None):
+          argsparse_groups=None, parts_to_load=None, logger=None,
+          useseparator=False):
     """ Function which initialize the application
 
     ::
@@ -28,6 +29,8 @@ def start(processName, version=release.version, prompt=PROMPT,
     :param argsparse_groups: list of the group of option for argparse
     :param parts_to_load: group of blok to load
     :param logger: option to configure  logging
+    :param useseparator: boolean, indicate if argsparse option are split
+        betwwen two application
     :rtype: registry if the database name is in the configuration
     """
     from .blok import BlokManager
@@ -42,7 +45,8 @@ def start(processName, version=release.version, prompt=PROMPT,
     if argsparse_groups is not None:
         ArgsParseManager.load(description=description,
                               argsparse_groups=argsparse_groups,
-                              parts_to_load=parts_to_load)
+                              parts_to_load=parts_to_load,
+                              useseparator=useseparator)
 
     if logger is None:
         logger = {}