+++ anyblok/bloks/io/tests/test_io_export_csv.py
@@ -6,7 +6,6 @@
 # v. 2.0. If a copy of the MPL was not distributed with this file,You can
 # obtain one at http://mozilla.org/MPL/2.0/.
 from anyblok.tests.testcase import BlokTestCase
-from anyblok.registry import RegistryManager
 
 
 class TestIOExportCSV(BlokTestCase):
@@ -14,7 +13,6 @@ class TestIOExportCSV(BlokTestCase):
     @classmethod
     def setUpClass(cls):
         super(TestIOExportCSV, cls).setUpClass()
-        RegistryManager.add_needed_bloks('io')
 
     def setUp(self):
         super(TestIOExportCSV, self).setUp()