+++ anyblok/test_bloks/test_blok1/__init__.py
@@ -15,3 +15,8 @@ class TestBlok(Blok):
     @classmethod
     def import_declaration_module(cls):
         from . import test  # noqa
+
+    @classmethod
+    def reload_declaration_module(cls, reload):
+        from . import test
+        reload(test)